Added Map class that handles map loading, rendering blocktypes etc. Also refactored a bunch of classes to work better. Fixed delta in Game class

+++ src/com/piratkopia13/pixelplanet/engine/core/Time.java
@@ -6,8 +6,8 @@ public class Time {
 
     private static double delta;
 
-    public static long getTime(){
-        return System.nanoTime();
+    public static double getTime(){
+        return (double)System.nanoTime();
     }
 
     public static double getDelta(){
